requirements.txt
# Core dependencies for Marketplace Bot
aiohttp>=3.13.1
flask>=3.1.2
python-dotenv>=1.1.1
python-telegram-bot[job-queue]>=22.5
werkzeug>=3.1.3

# Additional utilities
requests>=2.31.0
